Refactor iam.rb to file-per-method
