+++ google-cloud-storage/samples/storage_add_bucket_conditional_iam_binding.rb
@@ -1,0 +1,53 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+def add_bucket_conditional_iam_binding bucket_name:, role:, member:, title:, description:, expression:
+  # [START storage_add_bucket_conditional_iam_binding]
+  # bucket_name = "Your Google Cloud Storage bucket name"
+  # role        = "Bucket-level IAM role"
+  # member      = "Bucket-level IAM member"
+  # title       = "Condition Title"
+  # description = "Condition Description"
+  # expression  = "Condition Expression"
+
+  require "google/cloud/storage"
+
+  storage = Google::Cloud::Storage.new
+  bucket = storage.bucket bucket_name
+
+  bucket.policy requested_policy_version: 3 do |policy|
+    policy.version = 3
+    policy.bindings.insert(
+      role:      role,
+      members:   member,
+      condition: {
+        title:       title,
+        description: description,
+        expression:  expression
+      }
+    )
+  end
+
+  puts "Added #{member} with role #{role} to #{bucket_name} with condition #{title} #{description} #{expression}"
+  # [END storage_add_bucket_conditional_iam_binding]
+end
+
+if $PROGRAM_NAME == __FILE__
+  add_bucket_conditional_iam_binding bucket_name: ARGV.shift,
+                                     role:        ARGV.shift,
+                                     member:      ARGV.shift,
+                                     title:       ARGV.shift,
+                                     description: ARGV.shift,
+                                     expression:  ARGV.shift
+end
